separate schema from class
