+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/execution/ExecutionProof.java
@@ -1,3 +1,16 @@
+/*
+ * Copyright Consensys Software Inc., 2025
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
 package tech.pegasys.teku.spec.datastructures.execution;
 
 import tech.pegasys.teku.infrastructure.ssz.SszVector;
@@ -10,53 +23,36 @@ import tech.pegasys.teku.infrastructure.ssz.schema.SszPrimitiveSchemas;
 import tech.pegasys.teku.infrastructure.ssz.schema.SszVectorSchema;
 import tech.pegasys.teku.infrastructure.ssz.tree.TreeNode;
 
-public class ExecutionProof extends Container4<ExecutionProof, SszBytes32,  SszUInt64, SszUInt64, SszVector<SszByte>> {
+public class ExecutionProof
+    extends Container4<ExecutionProof, SszBytes32, SszUInt64, SszUInt64, SszVector<SszByte>> {
 
-    public static class ExecutionProofSchema extends ContainerSchema4<ExecutionProof, SszBytes32,  SszUInt64, SszUInt64, SszVector<SszByte>> {
-        public ExecutionProofSchema() {
-            super(
-                "ExecutionProof",
-                namedSchema("block_hash", SszPrimitiveSchemas.BYTES32_SCHEMA),
-                namedSchema("subnet_id", SszPrimitiveSchemas.UINT64_SCHEMA),
-                namedSchema("version", SszPrimitiveSchemas.UINT64_SCHEMA),
-                namedSchema("proof_data", SszVectorSchema.create(SszPrimitiveSchemas.BYTE_SCHEMA, 65536)) // Assuming max size of proof_data is 65536 bytes
-            );
-        }
+  private static final ExecutionProofSchema SSZ_SCHEMA = new ExecutionProofSchema();
 
-        @Override
-        public ExecutionProof createFromBackingNode(final TreeNode node) {
-            return new ExecutionProof(this, node);
-        }
+  public ExecutionProof(final ExecutionProofSchema type, final TreeNode node) {
+    super(type, node);
+  }
 
-    }
+  public ExecutionProof(
+      final SszBytes32 blockHash,
+      final SszUInt64 subnetId,
+      final SszUInt64 version,
+      final SszVector<SszByte> proofData) {
+    super(SSZ_SCHEMA, blockHash, subnetId, version, proofData);
+  }
 
-    public static final ExecutionProofSchema SSZ_SCHEMA = new ExecutionProofSchema();
+  public SszBytes32 getBlockHash() {
+    return getField0();
+  }
 
-    private ExecutionProof(final ExecutionProofSchema type, final TreeNode node) {
-        super(type, node);
-    }
+  public SszUInt64 getSubnetId() {
+    return getField1();
+  }
 
-    public ExecutionProof(
-        final SszBytes32 blockHash,
-        final SszUInt64 subnetId,
-        final SszUInt64 version,
-        final SszVector<SszByte> proofData) {
-        super(SSZ_SCHEMA, blockHash, subnetId, version, proofData);
-    }
+  public SszUInt64 getVersion() {
+    return getField2();
+  }
 
-
-    public SszBytes32 getBlockHash() {
-        return getField0();
-    }
-    public SszUInt64 getSubnetId() {
-        return getField1();
-    }
-
-    public SszUInt64 getVersion() {
-        return getField2();
-    }
-
-    public SszVector<SszByte> getProofData() {
-        return getField3();
-    }
+  public SszVector<SszByte> getProofData() {
+    return getField3();
+  }
 }
