+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/schemas/registry/SchemaTypes.java
@@ -53,6 +53,8 @@ import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayload;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeader;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadHeaderSchema;
 import tech.pegasys.teku.spec.datastructures.execution.ExecutionPayloadSchema;
+import tech.pegasys.teku.spec.datastructures.execution.ExecutionProof;
+import tech.pegasys.teku.spec.datastructures.execution.ExecutionProofSchema;
 import tech.pegasys.teku.spec.datastructures.execution.versions.capella.WithdrawalSchema;
 import tech.pegasys.teku.spec.datastructures.execution.versions.electra.ConsolidationRequestSchema;
 import tech.pegasys.teku.spec.datastructures.execution.versions.electra.DepositRequestSchema;
@@ -121,6 +123,8 @@ public class SchemaTypes {
       create("METADATA_MESSAGE_SCHEMA");
   public static final SchemaId<StatusMessageSchema<?>> STATUS_MESSAGE_SCHEMA =
       create("STATUS_MESSAGE_SCHEMA");
+  public static final SchemaId<ExecutionProofSchema> EXECUTION_PROOF_SCHEMA =
+      create("EXECUTION_PROOF_SCHEMA");
 
   // Altair
